A vowel plot function

This function will plot f1/f2 data. The function takes 4 obligatory arguments and optionally one can include a grouping variable ('group'). This function is a wrapper for ggplot. Columns from the data frame must be wrapped in quotation marks.

vowel_plot(data, vowel, f1, f2, group = NULL, print = FALSE)

Arguments

data

A data frame.

vowel

A factor with vowels as levels

f1

A continuous variable

f2

A continuous variable

group

An optional grouping variable for facets

print

Logical for printing descriptives

Examples

# Vowel plot without grouping variable

library("untidydata")
data("spanish_vowels")
#vowel_plot(data = spanish_vowels, vowel = 'vowel', f1 = 'f1', f2 = 'f2', 
#           group = NULL)

# Vowel plot with grouping variable
# vowel_plot(data = spanish_vowels, vowel = 'vowel', f1 = 'f1', f2 = 'f2', 
#            group = 'gender')